requirements.txt
click
tira>=0.0.194
python-terrier==0.13.0
ir-datasets
sentence-transformers
faiss-cpu
nltk